+++ src/clj_cbor/error.clj
@@ -14,4 +14,20 @@
   codec-exception!)
 
 
-; TODO: exception hierarchy
+
+;; ## Error Hierarchy
+
+;; Encoding errors.
+(derive :clj-cbor.header/negative-info-code ::encoding-error)
+(derive :clj-cbor.header/overflow-info-code ::encoding-error)
+(derive :clj-cbor.codec/illegal-simple-type ::encoding-error)
+(derive :clj-cbor.codec/unsupported-type    ::encoding-error)
+
+;; Decoding errors.
+(derive :clj-cbor.header/reserved-info-code ::decoding-error)
+(derive :clj-cbor.codec/illegal-chunk       ::decoding-error)
+(derive :clj-cbor.codec/illegal-stream      ::decoding-error)
+(derive :clj-cbor.codec/missing-map-value   ::decoding-error)
+(derive :clj-cbor.codec/duplicate-map-key   ::decoding-error)
+(derive :clj-cbor.codec/tag-handling-error  ::decoding-error)
+(derive :clj-cbor.codec/unexpected-break    ::decoding-error)