+++ src/clj_cbor/header.clj
@@ -1,7 +1,6 @@
 (ns clj-cbor.header
   "Functions for reading and writing CBOR headers."
   (:require
-    [clj-cbor.data.model :as data]
     [clj-cbor.error :as error]
     [clojure.string :as str])
   (:import
@@ -10,61 +9,78 @@
       DataOutputStream)))
 
 
+(def major-types
+  "Vector of major type keywords, indexed by the three-bit values 0-7. (§2.1)"
+  [:unsigned-integer
+   :negative-integer
+   :byte-string
+   :text-string
+   :data-array
+   :data-map
+   :tagged-value
+   :simple-value])
+
+
+(def ^:private major-type-codes
+  "Map of major type keywords to code values."
+  (zipmap major-types (range)))
+
+
+
 ;; ## Encoding Functions
 
-(let [mtype-codes (zipmap data/major-types (range))]
-  (defn write
-    "Writes a header byte for the given major-type and additional info numbers.
-    Returns the number of bytes written."
-    [^DataOutputStream out mtype info]
-    (let [header (-> (bit-and (mtype-codes mtype) 0x07)
-                     (bit-shift-left 5)
-                     (bit-or (bit-and (long info) 0x1F)))]
-      (.writeByte out header))
-    1))
+(defn write-leader
+  "Writes a header byte for the given major-type and additional info numbers.
+  Returns the number of bytes written."
+  [^DataOutputStream out mtype info]
+  (let [header (-> (bit-and (major-type-codes mtype) 0x07)
+                   (bit-shift-left 5)
+                   (bit-or (bit-and (long info) 0x1F)))]
+    (.writeByte out header))
+  1)
 
 
-(defn write-major-int
+(defn write
   "Writes a header byte for the given major-type, plus extra bytes to encode
-  the given integer value. Always writes the smallest possible representation.
+  the given integer code. Always writes the smallest possible representation.
   Returns the number of bytes written."
   [^DataOutputStream out mtype i]
   (cond
     (neg? i)
       (error/*handler*
-        ::negative-int-code
+        ::negative-info-code
         (str "Cannot write negative integer code: " i)
         {:code i})
     (<= i 23)
-      (do (write out mtype i)
+      (do (write-leader out mtype i)
           1)
     (<= i 0xFF)
-      (do (write out mtype 24)
+      (do (write-leader out mtype 24)
           (.writeByte out i)
           2)
     (<= i 0xFFFF)
-      (do (write out mtype 25)
+      (do (write-leader out mtype 25)
           (.writeShort out i)
           3)
     (<= i Integer/MAX_VALUE)
-      (do (write out mtype 26)
+      (do (write-leader out mtype 26)
           (.writeInt out i)
           5)
     (<= i 0xFFFFFFFF)
-      (do (write out mtype 26)
+      (do (write-leader out mtype 26)
           (.writeInt out (+ Integer/MIN_VALUE (- (dec i) Integer/MAX_VALUE)))
           5)
     (<= i Long/MAX_VALUE)
-      (do (write out mtype 27)
+      (do (write-leader out mtype 27)
           (.writeLong out i)
           9)
     (<= i (* -2N Long/MIN_VALUE))
-      (do (write out mtype 27)
+      (do (write-leader out mtype 27)
           (.writeLong out (+ Long/MIN_VALUE (- (dec i) Long/MAX_VALUE)))
           9)
     :else
       (error/*handler*
-        ::header-int-overflow
+        ::overflow-info-code
         (str "Cannot write integer code requiring 9 bytes of space: " i)
         {:code i})))
 
@@ -80,7 +96,7 @@
        (bit-and 0xE0)
        (bit-shift-right 5)
        (bit-and 0x07)
-       (data/major-types))
+       (major-types))
    (bit-and header 0x1F)])
 
 
@@ -107,7 +123,7 @@
       value)))
 
 
-(defn read-size
+(defn read-code
   "Reads a size value from the initial bytes of the input stream. Returns
   either a number, the keyword `:indefinite`, or calls the error handler on
   reserved info codes."
@@ -123,7 +139,7 @@
       27 (read-unsigned-long input)
       (28 29 30)
         (error/*handler*
-          ::reserved-length
+          ::reserved-info-code
           (format "Additional information int code %d is reserved."
                   info)
           {:info info})