+++ src/clj_cbor/tags/text.clj
@@ -2,9 +2,11 @@
   "Built-in tag support for the text extensions in RFC 7049. See section
   2.4.4."
   (:require
-    [clj-cbor.data.model :as data])
+    [clj-cbor.data.core :as data])
   (:import
     java.net.URI
+    java.nio.ByteBuffer
+    java.util.UUID
     java.util.regex.Pattern))
 
 
@@ -47,6 +49,23 @@
 ; tag 37
 ; https://github.com/lucas-clemente/cbor-specs/blob/master/uuid.md
 
+(defn format-uuid
+  [^UUID value]
+  (let [data (ByteBuffer/allocate 16)]
+    (.putLong data (.getMostSignificantBits value))
+    (.putLong data (.getLeastSignificantBits value))
+    (data/tagged-value 37 (.array data))))
+
+
+(defn parse-uuid
+  [tag value]
+  (when-not (data/bytes? value)
+    (throw (ex-info (str "UUIDs must be tagged byte strings, got: "
+                         (class value))
+                    {:tag tag, :value value})))
+  (let [data (ByteBuffer/wrap value)]
+    (UUID. (.getLong data) (.getLong data))))
+
 
 
 ;; ## Codec Formatter/Handler Maps
@@ -54,10 +73,12 @@
 (def text-write-handlers
   "Map of text types to formatting functions."
   {URI     format-uri
+   UUID    format-uuid
    Pattern format-pattern})
 
 
 (def text-read-handlers
   "Map of tag handlers to parse text values."
   {32 parse-uri
-   35 parse-pattern})
+   35 parse-pattern
+   37 parse-uuid})